agent/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
google-generativeai
langchain
langchain-google-genai
pymongo